Resume Screener/requirements.txt
python-docx==1.0.1
PyPDF2==3.0.1
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
ratelimit==2.2.1
colorama==0.4.6
tqdm==4.66.1
google-generativeai==0.4.0
openai==1.16.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-cov==4.1.0